Create request.go and move request struct into it
